refactor supabase client
