+++ apps/expo/src/utils/supabase.ts
@@ -1,0 +1,27 @@
+import Constants from "expo-constants";
+import * as SecureStore from "expo-secure-store";
+import { createClient } from "@supabase/supabase-js";
+
+import "react-native-url-polyfill/auto";
+
+const ExpoSecureStoreAdapter = {
+  getItem: (key: string) => SecureStore.getItemAsync(key),
+  setItem: (key: string, value: string) => SecureStore.setItemAsync(key, value),
+  removeItem: (key: string) => SecureStore.deleteItemAsync(key),
+};
+
+export const supabase = createClient(
+  // App Throws if these are not defined, so we can safely cast
+  Constants.expoConfig?.extra?.SUPABASE_URL as string,
+  Constants.expoConfig?.extra?.SUPABASE_ANON_KEY as string,
+  {
+    auth: {
+      storage: ExpoSecureStoreAdapter,
+      autoRefreshToken: true,
+      persistSession: true,
+      detectSessionInUrl: true,
+      // @ts-expect-error - FIXME: Do I need a canary???
+      flowType: "pkce",
+    },
+  },
+);
